backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
google-genai>=1.0
pdfplumber>=0.11
python-multipart>=0.0.9
python-dotenv>=1.0
pydantic>=2.0
